requirements.txt
discord.py
openai
httpx
requests
flask
line-bot-sdk